Add tree editor example
Adds a new template `Tree Editor` to the generator which allows to generate an extension with a tree master detail editor.
Additionally, add a new menu Tree Editor with an entry to generate an example file to open in the editor.
The tree editor uses a coffee machine example model.

Co-authored-by: Florian Gareis <mail@zoker.me>
Co-authored-by: Jonas Helming <jhelming@eclipsesource.com>
Signed-off-by: Lucas Koehler <lkoehler@eclipsesource.com>